Moved interfaces and base classes into separate files and placed into corresponding folders
