+++ DungeonMaster.Data/Models/BaseModel/BaseModel.cs
@@ -1,0 +1,6 @@
+﻿namespace DungeonMaster.Data.Models
+{
+    public abstract class BaseModel
+    {
+    }
+}
