+++ DungeonMaster.Data/Models/BaseModel/IModel.cs
@@ -1,0 +1,7 @@
+﻿namespace DungeonMaster.Data.Models
+{
+    public interface IModel
+    {
+        int Id { get; set; }
+    }
+}
