+++ DungeonMaster.Data/Models/BaseModel/Model.cs
@@ -5,13 +5,4 @@
     {
         public virtual int Id { get; set; }
     }
-
-    public abstract class BaseModel
-    {
-    }
-
-    public interface IModel
-    {
-        int Id { get; set; }
-    }
 }
