+++ DungeonMaster.Repositories/BaseRepository/IGenericRepository.cs
@@ -1,0 +1,17 @@
+﻿namespace DungeonMaster.Repositories
+{
+    using DungeonMaster.Data.Models;
+    using System;
+    using System.Collections.Generic;
+    using System.Linq.Expressions;
+
+    public interface IGenericRepository<T> where T : BaseModel
+    {
+        IEnumerable<T> GetAll();
+        IEnumerable<T> FindBy(Expression<Func<T, bool>> predicate);
+        T Add(T entity);
+        T Delete(T entity);
+        void Edit(T entity);
+        void Save();
+    }
+}
