+++ DungeonMaster.Repositories/CharacterRepository/ICharacterRepository.cs
@@ -1,0 +1,9 @@
+﻿namespace DungeonMaster.Repositories
+{
+    using DungeonMaster.Data.Models;
+
+    public interface ICharacterRepository : IGenericRepository<Character>
+    {
+        Character GetById(int id);
+    }
+}
