+++ DungeonMaster.Repositories/ClassRepository/IClassRepository.cs
