+++ DungeonMaster.Repositories/UnitOfWork/IUnitOfWork.cs
@@ -1,0 +1,9 @@
+﻿namespace DungeonMaster.Repositories
+{
+    using System;
+
+    public interface IUnitOfWork : IDisposable
+    {
+        int Commit();
+    }
+}
