+++ DungeonMaster.Services/BaseService/IModelService.cs
@@ -1,0 +1,13 @@
+﻿namespace DungeonMaster.Services
+{
+    using DungeonMaster.Data.Models;
+    using System.Collections.Generic;
+
+    public interface IModelService<T> : IService where T : BaseModel
+    {
+        void Create(T entity);
+        void Delete(T entity);
+        IEnumerable<T> GetAll();
+        void Update(T entity);
+    }
+}
