+++ DungeonMaster.Services/BaseService/IService.cs
@@ -1,0 +1,6 @@
+﻿namespace DungeonMaster.Services
+{
+    public interface IService
+    {
+    }
+}
