+++ DungeonMaster.Services/CharacterService/ICharacterService.cs
@@ -1,0 +1,9 @@
+﻿namespace DungeonMaster.Services
+{
+    using DungeonMaster.Data.Models;
+
+    public interface ICharacterService : IModelService<Character>
+    {
+        Character GetById(int id);
+    }
+}
