--- DungeonMaster.Services/ICharacterService.cs
@@ -1,9 +1,0 @@
-﻿using DungeonMaster.Data.Models;
-
-namespace DungeonMaster.Services
-{
-    public interface ICharacterService
-    {
-        Character CreateCharacter(Character scaffoldedCharacter);
-    }
-}
